+++ test/test-util/jest-setup.js
@@ -1,0 +1,7 @@
+'use strict';
+
+require('./raf');
+const Enzyme = require('enzyme');
+const Adapter = require('enzyme-adapter-react-16');
+
+Enzyme.configure({ adapter: new Adapter() });