+++ CHANGELOG.md
@@ -9,7 +9,8 @@
 - [Fix] Fragments in the URL take precedence over saved scroll positions in the history.
 ## 0.11.3
 
-- [Fix] Allow for React 16 as peer dependency.
+- [Chore] Allow for React 16 as peer dependency.
+- [Chore] Update all other dependencies.
 
 ## 0.11.2